Fixed turing errors

+++ 3_2_Pocet_binarnych_vyhladavacich_stromov/3_2_Pocet_binarnych_vyhladavacich_stromov/count.hpp
@@ -1,6 +1,6 @@
 #ifndef count_hpp
 #define count_hpp
 
-unsigned numberOfTrees(int count);
+unsigned long numberOfTrees(int count);
 
 #endif
